Add new issue template for doc bugs (#12)
This change adds an issue template for reporting on gaps in
documentation. It is simpler than either the bug report or feature
request template. Default labels were added  to this template and
to the existing issue templates.

This change also moves the pull request template directly into the
.github/ folder. It was not being picked up without specifying a query
param in a URL. This should now be the default location.

Resolves: #14, #15